# Speed up reading Hive and Drill partitions