@ig248 2020-01-02
In fact, this change could/should also be propagated to https://github.com/dask/dask/blob/c69d10f97cadd06f147817f0084bddeadb8195e6/dask/dataframe/io/parquet/fastparquet.py#L30