@martindurant 2020-01-02
At a quick glance, this looks like an excellent change. I have not had change to look in detail, so will leave it open a little while longer in case anyone else has something to comment.