@ig248 2020-01-22
@martindurant do you have any thoughts/suggestions? The only thing I can think of is to add atomic unit tests for the functions added in this PR, if you prefer.  It's not urgent but I'd like to be able to install from `master`...