@rjzamora 2020-01-22
This looks great - Thanks @ig248 !

The parquet-related test suite is passing for me on my local machine with these changes.  I completely agree that the `paths_to_cats` changes should definitely be propagated to the copied code in dask :)