@martindurant 2020-01-22
Thanks for the approval, @rjzamora - I'll merge now. Hopefully someone has the chance to make the changes in Dask sometime.