@ig248 2020-01-22
Thanks @rjzamora and @martindurant, I can look at propagating the change to `dask`